requirements.txt
albumentations==2.0.4
fiftyone==1.3.0
matplotlib==3.10.0
numpy==2.2.3
optuna==4.0.0
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
scipy==1.15.1
shap==0.46.0
torch==2.4.1
torchcam==0.4.0
torchvision==0.19.1
tqdm==4.66.5
